+++ LC1/src/main/java/tv/piratemedia/lightcontroler/controller.java
@@ -42,7 +42,6 @@ import android.support.v4.widget.DrawerLayout;
 import android.support.v7.widget.Toolbar;
 import android.text.InputType;
 import android.util.DisplayMetrics;
-import android.util.Log;
 import android.util.TypedValue;
 import android.view.Display;
 import android.view.LayoutInflater;
@@ -81,6 +80,8 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
 
+import tv.piratemedia.lightcontroler.wear.DataLayerListenerService;
+
 public class controller extends ActionBarActivity {
 
     /**
@@ -93,8 +94,7 @@ public class controller extends ActionBarActivity {
     private static boolean micStarted = false;
     private static boolean candleMode = false;
     private static Context ctx;
-
-    private SaveState appState = null;
+    public SaveState appState = null;
     private static SharedPreferences prefs;
 
     private Toolbar mActionBarToolbar;
@@ -134,6 +134,7 @@ public class controller extends ActionBarActivity {
 
         appState = new SaveState(this);
         Utils = new utils(this);
+        new DataLayerListenerService();
     }
 
     class MyHandler extends Handler {
@@ -728,7 +729,8 @@ public class controller extends ActionBarActivity {
                 final View rootView = inflater.inflate(R.layout.rgbw_control, container, false);
 
                 SeekBar brightness = (SeekBar) rootView.findViewById(R.id.brightness);
-                ToggleButton io = (ToggleButton) rootView.findViewById(R.id.onoff);
+                Button on = (Button) rootView.findViewById(R.id.on);
+                Button off = (Button) rootView.findViewById(R.id.off);
                 Button disco = (Button) rootView.findViewById(R.id.disco);
                 Button dplus = (Button) rootView.findViewById(R.id.dplus);
                 Button dminus = (Button) rootView.findViewById(R.id.dminus);
@@ -748,7 +750,7 @@ public class controller extends ActionBarActivity {
                 modeSpinner.setAdapter(modeAdapter);
 
                 //Return State
-                io.setChecked(((controller)getActivity()).appState.getOnOff(getArguments().getInt(ARG_SECTION_NUMBER)));
+                //io.setChecked(((controller)getActivity()).appState.getOnOff(getArguments().getInt(ARG_SECTION_NUMBER)));
                 brightness.setProgress(((controller)getActivity()).appState.getBrightness(getArguments().getInt(ARG_SECTION_NUMBER)));
                 int savedColor = ((controller)getActivity()).appState.getColor(getArguments().getInt(ARG_SECTION_NUMBER));
                 if(savedColor < 0) {
@@ -794,8 +796,8 @@ public class controller extends ActionBarActivity {
                         if(!disabled) {
                             Controller.setColor(getArguments().getInt(ARG_SECTION_NUMBER), i);
                             ((controller) getActivity()).setActionbarColor(i);
-                            ToggleButton io = (ToggleButton) rootView.findViewById(R.id.onoff);
-                            io.setChecked(true);
+                            /*ToggleButton io = (ToggleButton) rootView.findViewById(R.id.onoff);
+                            io.setChecked(true);*/
                         }
                     }
                 });
@@ -804,8 +806,8 @@ public class controller extends ActionBarActivity {
                     @Override
                     public void onProgressChanged(SeekBar seekBar, int progress, boolean fromUser) {
                         Controller.setBrightness(getArguments().getInt(ARG_SECTION_NUMBER), progress);
-                        ToggleButton io = (ToggleButton) rootView.findViewById(R.id.onoff);
-                        io.setChecked(true);
+                        /*ToggleButton io = (ToggleButton) rootView.findViewById(R.id.onoff);
+                        io.setChecked(true);*/
                     }
 
                     @Override
@@ -819,7 +821,19 @@ public class controller extends ActionBarActivity {
                     }
                 });
 
-                io.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
+                on.setOnClickListener(new View.OnClickListener() {
+                    @Override
+                    public void onClick(View v) {
+                        Controller.LightsOn(getArguments().getInt(ARG_SECTION_NUMBER));
+                    }
+                });
+                off.setOnClickListener(new View.OnClickListener() {
+                    @Override
+                    public void onClick(View v) {
+                        Controller.LightsOff(getArguments().getInt(ARG_SECTION_NUMBER));
+                    }
+                });
+                /*io.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
                     @Override
                     public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
                         if (isChecked) {
@@ -828,14 +842,14 @@ public class controller extends ActionBarActivity {
                             Controller.LightsOff(getArguments().getInt(ARG_SECTION_NUMBER));
                         }
                     }
-                });
+                });*/
 
                 disco.setOnClickListener(new View.OnClickListener() {
                     @Override
                     public void onClick(View v) {
                         Controller.toggleDiscoMode(getArguments().getInt(ARG_SECTION_NUMBER));
-                        ToggleButton io = (ToggleButton) rootView.findViewById(R.id.onoff);
-                        io.setChecked(true);
+                        /*ToggleButton io = (ToggleButton) rootView.findViewById(R.id.onoff);
+                        io.setChecked(true);*/
                     }
                 });
 
@@ -843,8 +857,8 @@ public class controller extends ActionBarActivity {
                     @Override
                     public void onClick(View v) {
                         Controller.discoModeFaster();
-                        ToggleButton io = (ToggleButton) rootView.findViewById(R.id.onoff);
-                        io.setChecked(true);
+                        /*ToggleButton io = (ToggleButton) rootView.findViewById(R.id.onoff);
+                        io.setChecked(true);*/
                     }
                 });
 
@@ -852,8 +866,8 @@ public class controller extends ActionBarActivity {
                     @Override
                     public void onClick(View v) {
                         Controller.discoModeSlower();
-                        ToggleButton io = (ToggleButton) rootView.findViewById(R.id.onoff);
-                        io.setChecked(true);
+                        /*ToggleButton io = (ToggleButton) rootView.findViewById(R.id.onoff);
+                        io.setChecked(true);*/
                     }
                 });
 
@@ -861,8 +875,8 @@ public class controller extends ActionBarActivity {
                     @Override
                     public void onClick(View v) {
                         Controller.setToWhite(getArguments().getInt(ARG_SECTION_NUMBER));
-                        ToggleButton io = (ToggleButton) rootView.findViewById(R.id.onoff);
-                        io.setChecked(true);
+                        /*ToggleButton io = (ToggleButton) rootView.findViewById(R.id.onoff);
+                        io.setChecked(true);*/
                         ((controller) getActivity()).setActionbarColor(Color.parseColor("#ffee58"));
                     }
                 });
@@ -979,12 +993,13 @@ public class controller extends ActionBarActivity {
                 CircularSeekBar warmth = (CircularSeekBar) rootView.findViewById(R.id.warmth);
                 final TextView brightnessvalue = (TextView) rootView.findViewById(R.id.brightnessvalue);
                 final TextView warmthvalue = (TextView) rootView.findViewById(R.id.warmthvalue);
-                ToggleButton io = (ToggleButton) rootView.findViewById(R.id.onoff);
+                Button on = (Button) rootView.findViewById(R.id.on);
+                Button off = (Button) rootView.findViewById(R.id.off);
                 Button full = (Button) rootView.findViewById(R.id.full);
                 Button night = (Button) rootView.findViewById(R.id.night);
 
                 //Return State
-                io.setChecked(((controller)getActivity()).appState.getOnOff(getArguments().getInt(ARG_SECTION_NUMBER)));
+                //io.setChecked(((controller)getActivity()).appState.getOnOff(getArguments().getInt(ARG_SECTION_NUMBER)));
                 brightness.setProgress(((controller)getActivity()).appState.getBrightness(getArguments().getInt(ARG_SECTION_NUMBER)));
                 int savedColor = ((controller)getActivity()).appState.getColor(getArguments().getInt(ARG_SECTION_NUMBER));
                 if(savedColor < 0) {
@@ -1082,7 +1097,19 @@ public class controller extends ActionBarActivity {
                     }
                 });
 
-                io.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
+                on.setOnClickListener(new View.OnClickListener() {
+                    @Override
+                    public void onClick(View v) {
+                        Controller.LightsOn(getArguments().getInt(ARG_SECTION_NUMBER));
+                    }
+                });
+                off.setOnClickListener(new View.OnClickListener() {
+                    @Override
+                    public void onClick(View v) {
+                        Controller.LightsOff(getArguments().getInt(ARG_SECTION_NUMBER));
+                    }
+                });
+                /*io.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
                     @Override
                     public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
                         if (isChecked) {
@@ -1091,14 +1118,14 @@ public class controller extends ActionBarActivity {
                             Controller.LightsOff(getArguments().getInt(ARG_SECTION_NUMBER));
                         }
                     }
-                });
+                });*/
 
                 full.setOnClickListener(new View.OnClickListener() {
                     @Override
                     public void onClick(View v) {
                         Controller.setToFull(getArguments().getInt(ARG_SECTION_NUMBER));
-                        ToggleButton io = (ToggleButton) rootView.findViewById(R.id.onoff);
-                        io.setChecked(true);
+                        /*ToggleButton io = (ToggleButton) rootView.findViewById(R.id.onoff);
+                        io.setChecked(true);*/
                     }
                 });
 
@@ -1106,8 +1133,8 @@ public class controller extends ActionBarActivity {
                     @Override
                     public void onClick(View v) {
                         Controller.setToNight(getArguments().getInt(ARG_SECTION_NUMBER));
-                        ToggleButton io = (ToggleButton) rootView.findViewById(R.id.onoff);
-                        io.setChecked(true);
+                        /*ToggleButton io = (ToggleButton) rootView.findViewById(R.id.onoff);
+                        io.setChecked(true);*/
                     }
                 });