+++ LC1/src/main/java/tv/piratemedia/lightcontroler/utils.java
@@ -1,11 +1,24 @@
 package tv.piratemedia.lightcontroler;
 
+import android.app.Activity;
 import android.content.Context;
+import android.content.Intent;
 import android.net.ConnectivityManager;
 import android.net.DhcpInfo;
 import android.net.NetworkInfo;
 import android.net.wifi.WifiManager;
+import android.os.Bundle;
+import android.os.UserHandle;
 import android.util.Log;
+import android.util.Patterns;
+
+import com.google.android.gms.common.api.GoogleApiClient;
+import com.google.android.gms.common.api.PendingResult;
+import com.google.android.gms.common.api.ResultCallback;
+import com.google.android.gms.wearable.MessageApi;
+import com.google.android.gms.wearable.Node;
+import com.google.android.gms.wearable.NodeApi;
+import com.google.android.gms.wearable.Wearable;
 
 import org.apache.http.conn.util.InetAddressUtils;
 
@@ -15,17 +28,18 @@ import java.net.InterfaceAddress;
 import java.net.NetworkInterface;
 import java.net.SocketException;
 import java.net.UnknownHostException;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.Enumeration;
+import java.util.HashSet;
 import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import java.util.regex.PatternSyntaxException;
-
 /**
  * Created by eliotstocker on 15/11/14.
  */
-public class utils {
+public class utils extends Activity implements GoogleApiClient.ConnectionCallbacks  {
     public final static int IP_ADDRESS = 0;
     public final static int BROADCAST_ADDRESS = 1;
 
@@ -47,7 +61,6 @@ public class utils {
             WifiManager wifi = (WifiManager) mCtx.getSystemService(Context.WIFI_SERVICE);
             DhcpInfo dhcp = wifi.getDhcpInfo();
             // handle null somehow
-
             switch(type) {
                 case BROADCAST_ADDRESS :
                     Log.d("Utils", "get Broadcast");
@@ -84,8 +97,7 @@ public class utils {
 
     public String getWifiName() {
         WifiManager wifi = (WifiManager) mCtx.getSystemService(Context.WIFI_SERVICE);
-        String SSID = wifi.getConnectionInfo().getSSID();
-        return SSID.replace("\"","");
+        return wifi.getConnectionInfo().getSSID().substring(1, wifi.getConnectionInfo().getSSID().length() - 1);
     }
 
     public String GetWifiMac() {
@@ -107,6 +119,10 @@ public class utils {
         }
     }
 
+    public static boolean validHost(String url) {
+        return Patterns.WEB_URL.matcher(url).matches();
+    }
+
 
     public boolean validMac(String MAC) {
         if (MAC == null || MAC.isEmpty()) return false;
@@ -120,5 +136,15 @@ public class utils {
         } catch (PatternSyntaxException ex) {
             return false;
         }
+    }
+
+    public void onConnected(Bundle bundle) {
+
+    }
+
+
+    @Override
+    public void onConnectionSuspended(int i) {
+
     }
 }