+++ changelog.md
@@ -6,6 +6,7 @@ updated 16th Nov 2014
 
 1.0
 ---
+Initial Release
 
 1.5
 ---
@@ -16,10 +17,23 @@ updated 16th Nov 2014
 * New Icon
 * App color changes to that of bulbs
 
-master
-------
+1.6
+---
 * Save state locally so that app remembers bulb color/brightness/io settings
 * add service discovery, to find control boxes
 * add wifi setup for new/rest boxes
 * add touch to select color
-* serveral stability and architechural improvements
+* Several stability and architectural improvements
+* dual white LED support
+
+1.6.1
+-----
+* Fix first run UI issue
+
+1.7
+---
+* Android Wear Application
+* Widget Improvements
+* Full API to implement light control into 3rd part apps
+* Allow Host for Control Box instead of IP
+